refactor(app): move protocol visualization components under pages to organisms
move protocol visualization components under pages to organisms

close AUTH-2521
