+++ app/src/molecules/SlotDetailsEmptyState/slotdetailsemptystate.module.css
@@ -1,0 +1,17 @@
+.slot_details_active_step {
+  display: flex;
+  flex-direction: column;
+  padding: 0 var(--spacing-12) var(--spacing-12) var(--spacing-12);
+  gap: var(--spacing-8);
+}
+
+.empty_state_box_container {
+  padding: 0 var(--spacing-12);
+}
+
+.empty_state_box {
+  width: 100%;
+  height: 8.330625rem;
+  border-radius: var(--border-radius-4);
+  background-color: var(--grey-10);
+}
