+++ app/src/organisms/Desktop/ProtocolVisualization/DeckView/DeckViewOverlay/deckviewoverlay.module.css
@@ -1,0 +1,18 @@
+.deck_overlay {
+  display: flex;
+  width: 100%;
+  height: 100%;
+  align-items: flex-end;
+  justify-content: center;
+  border-radius: var(--border-radius-4);
+  background-color: transparent;
+  gap: var(--spacing-8);
+}
+
+.text_background {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+  border-radius: 0 0 var(--border-radius-4) var(--border-radius-4);
+  background-color: var(--slot-fill-color);
+}
