+++ app/src/organisms/Desktop/ProtocolVisualization/DeckView/Ot2FixedTrashCommandSummary.tsx
