+++ app/src/organisms/Desktop/ProtocolVisualization/ModuleSlotDetails/modulesslotdetails.module.css
@@ -1,0 +1,31 @@
+.module_details_status_container {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  padding: var(--spacing-16);
+
+  /* border-radius: var(--border-radius-8); */
+  background-color: var(--white);
+  gap: var(--spacing-4);
+}
+
+.module_details_status {
+  display: flex;
+  justify-content: space-between;
+  padding-left: var(--spacing-16);
+}
+
+.detail_container {
+  display: flex;
+  height: 100%;
+  flex-direction: column;
+  border-radius: var(--border-radius-8);
+  background-color: var(--white);
+}
+
+.slot_details_active_step {
+  display: flex;
+  flex-direction: column;
+  padding: 0 var(--spacing-12) var(--spacing-12) var(--spacing-12);
+  gap: var(--spacing-8);
+}
