+++ app/src/organisms/Desktop/ProtocolVisualization/SlotDetails/index.tsx
@@ -2,11 +2,11 @@ import { Divider, RobotInfoLabel, StyledText } from '@opentrons/components'
 import { getModuleDeckLabel } from '@opentrons/shared-data'
 import { getFullStackFromLabwares } from '@opentrons/step-generation'
 
-import { LabwareSlotContainer } from '/app/organisms/Desktop/ProtocolVisualization/LabwareSlotContainer'
+import { SlotDetailsEmptyState } from '/app/molecules/SlotDetailsEmptyState'
 
-import { ModuleSlotDetails } from './ModuleSlotDetails'
-import { SlotDetailsEmptyState } from './SlotDetailsEmptyState'
-import { TrashSlotDetails } from './TrashSlotDetails'
+import { LabwareSlotContainer } from '../LabwareSlotContainer'
+import { ModuleSlotDetails } from '../ModuleSlotDetails'
+import { TrashSlotDetails } from '../TrashSlotDetails'
 import styles from './visualization.module.css'
 
 import type {
