+++ app/src/organisms/Desktop/ProtocolVisualization/StepDetailContainer.tsx
@@ -6,11 +6,9 @@ import { PipetteContainer } from './PipetteContainer'
 import styles from './stepdetailcontainer.module.css'
 import { TipDisposalContainer } from './TipDisposalContainer'
 import { TipPickupContainer } from './TipPickupContainer'
-import {
-  getActiveSlotForLabwareDetails,
-  getActiveSlotForTiprackDetails,
-  getIsPipetteActive,
-} from './utils'
+import { getActiveSlotForLabwareDetails } from './utils/getActiveSlotForLabwareDetails'
+import { getActiveSlotForTiprackDetails } from './utils/getActiveSlotForTiprackDetails'
+import { getIsPipetteActive } from './utils/getIsPipetteActive'
 
 import type { Liquid, RunTimeCommand } from '@opentrons/shared-data'
 import type { InvariantContext, RobotState } from '@opentrons/step-generation'
