+++ app/src/organisms/Desktop/ProtocolVisualization/VisualizerContainer/index.tsx
@@ -10,13 +10,13 @@ import {
   getResultingTimelineFrameFromRunCommands,
 } from '@opentrons/step-generation'
 
+import { CommandSteps } from '/app/organisms/Desktop/ProtocolVisualization/CommandSteps'
 import { Controls } from '/app/organisms/Desktop/ProtocolVisualization/Controls'
+import { DeckView } from '/app/organisms/Desktop/ProtocolVisualization/DeckView'
 import { StepDetailContainer } from '/app/organisms/Desktop/ProtocolVisualization/StepDetailContainer'
 import { stepDetailViewerOpenAction } from '/app/redux/shell'
 import { getProtocolDisplayName } from '/app/transformations/protocols'
 
-import { CommandSteps } from './CommandSteps'
-import { DeckView } from './DeckView'
 import styles from './visualizercontainer.module.css'
 
 import type { MouseEvent } from 'react'
